refact: Separating all components
