+++ src/Button.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Button({ children, onClick }) {
+  return (
+    <button type="button" onClick={onClick}>
+      {children}
+    </button>
+  );
+}
