+++ src/Title.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Title() {
+  return (
+    <p>
+      Counter
+    </p>
+  );
+}
